+++ src/cli/show.rs
@@ -1,6 +1,7 @@
 use crate::{cli::Subcommand, util::Directory, GlobalArgs};
 use async_trait::async_trait;
 use clap::Parser;
+use std::process::ExitCode;
 
 /// Show meta information about slumber
 #[derive(Clone, Debug, Parser)]
@@ -17,10 +18,10 @@ enum ShowTarget {
 
 #[async_trait]
 impl Subcommand for ShowCommand {
-    async fn execute(self, _global: GlobalArgs) -> anyhow::Result<()> {
+    async fn execute(self, _global: GlobalArgs) -> anyhow::Result<ExitCode> {
         match self.target {
             ShowTarget::Dir => println!("{}", Directory::root()),
         }
-        Ok(())
+        Ok(ExitCode::SUCCESS)
     }
 }